requirements.txt
langchain>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0
chromadb>=0.5.0
langchain-chroma>=1.1.0
python-dotenv>=1.0.0
numpy>=1.26.0
